Update 唐怡婷1.py to optimize the nearest neighbor search algorithm
Through the above comment markings, you can clearly see that a while loop is added to the main function, which is used to obtain the coordinates of the target point through multiple interactions, and then perform the nearest neighbor search and visualization display accordingly. The program will terminate until the user enters non-numeric values, thus achieving a more flexible function that is different from the original code.